refactor for release
