+++ hydromt_delft3dfm/workflows/__init__.py
@@ -4,7 +4,6 @@ from .boundaries import *
 from .branches import *
 from .crosssections import *
 from .dem import *
-from .helper import *
 from .manholes import *
 from .mesh import *
 from .region import *
